+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added `QasGallery`
 
+### Fixed
+
+- Added cloned value for `field` in computed `children` inside `QasNestedFields`
+
 ## 2.2.0 - 2021-05-26
 
 ### Added